requirements.txt
requests
python-dotenv
arize-phoenix>=3.0.0
openinference-instrumentation-openai>=0.1.1
openai>=1.0.0

